# pkg/api/docs/swagger.yaml
basePath: /
definitions:
...
  api.TokenValidationResponse:
    properties:
      expires_at:
        type: string
      token_name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Source Code
    url: https://github.com/stefanprodan/podinfo
  description: Go microservice template for Kubernetes.
  license:
    name: MIT License
    url: https://github.com/stefanprodan/podinfo/blob/master/LICENSE
  title: Podinfo API
  version: "2.0"
paths:
...
  /cache/{key}:
    delete:
      consumes:
      - application/json
      description: deletes the key and its value from cache
      parameters:
      - description: Key to delete
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
      summary: Delete payload from cache
      tags:
      - HTTP API
    get:
      consumes:
      - application/json
      description: returns the content from cache if key exists
      parameters:
      - description: Key to load from cache
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get payload from cache
      tags:
      - HTTP API
    post:
      consumes:
      - application/json
      description: writes the posted content in cache
      parameters:
      - description: Key to save to
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
      summary: Save payload in cache
      tags:
      - HTTP API
...
  /store/{hash}:
    get:
      consumes:
      - application/json
      description: returns the content of the file /data/hash if exists
      parameters:
      - description: hash value
        in: path
        name: hash
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: file
          schema:
            type: string
      summary: Download file
      tags:
      - HTTP API
...
  /token/validate:
    post:
      consumes:
      - application/json
      description: validates the JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TokenValidationResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Validate JWT token
      tags:
      - HTTP API
...
schemes:
- http
- https
swagger: "2.0"